Add files via upload
This file specifically looks at Yosemite Wilderness trailhead permits that include the proper vector (direction arrow) that allow an overnight attempt with a Lyell/McLure summit in that area of the park. Note this does NOT permit exiting the park or any overnight outside the park, it's just to get you on the trail and overnight in the park somewhere other than your car in Tuolumne valley for a Lyell or similar summit attempt.